Reqire token for all dashboard endpoints
