+++ .gitignore
@@ -135,3 +135,6 @@ dmypy.json
 
 #dev db
 *.db
+
+# Constants
+config.py
