+++ app/api/auth_utils.py
@@ -1,8 +1,25 @@
+from datetime import datetime, timedelta
+from typing import Union
+
 from fastapi.security import OAuth2PasswordBearer
 from passlib.context import CryptContext
+from jose import jwt
+
+from config import SECRET_KEY, ALGORITHM
 
 password_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
 oauth2_scheme = OAuth2PasswordBearer(tokenUrl="auth/login", scheme_name="JWT")
+
+
+def create_access_token(data: dict, expires_delta: Union[timedelta, None] = None):
+    to_encode = data.copy()
+    if expires_delta:
+        expire = datetime.utcnow() + expires_delta
+    else:
+        expire = datetime.utcnow() + timedelta(minutes=15)
+    to_encode.update({"exp": expire})
+    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
+    return encoded_jwt
 
 
 def get_hashed_password(password: str) -> str:
