+++ app/api/routes/auth.py
@@ -1,30 +1,23 @@
-from datetime import datetime, timedelta
-from typing import Optional, Union
+from datetime import timedelta
+from typing import Optional
 
 from fastapi import APIRouter, status, Depends, HTTPException
 from fastapi.security import OAuth2PasswordRequestForm
-from jose import JWTError, jwt
 
 from pydantic import BaseModel
 
-from app.api.auth_utils import verify_password, oauth2_scheme
+from app.api.auth_utils import verify_password, create_access_token
 from app.crud.user import UserCRUD
+from app.dependencies.auth import get_current_user
 from app.models.user import UserCreate, User
+from config import ACCESS_TOKEN_EXPIRE_MINUTES
 
 router = APIRouter()
-
-SECRET_KEY = "temp"
-ALGORITHM = "HS256"
-ACCESS_TOKEN_EXPIRE_MINUTES = 30
 
 
 class Token(BaseModel):
     access_token: str
     token_type: str
-
-
-class TokenData(BaseModel):
-    username: str | None = None
 
 
 async def authenticate_user(username: str, password: str, user_crud: UserCRUD) -> Optional[User]:
@@ -33,37 +26,6 @@ async def authenticate_user(username: str, password: str, user_crud: UserCRUD) -
         return None
     if not verify_password(password, user.hashed_password):
         return None
-    return user
-
-
-def create_access_token(data: dict, expires_delta: Union[timedelta, None] = None):
-    to_encode = data.copy()
-    if expires_delta:
-        expire = datetime.utcnow() + expires_delta
-    else:
-        expire = datetime.utcnow() + timedelta(minutes=15)
-    to_encode.update({"exp": expire})
-    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
-    return encoded_jwt
-
-
-async def get_current_user(token: str = Depends(oauth2_scheme), user_crud: UserCRUD = Depends(UserCRUD)) -> User:
-    credentials_exception = HTTPException(
-        status_code=status.HTTP_401_UNAUTHORIZED,
-        detail="Could not validate credentials",
-        headers={"WWW-Authenticate": "Bearer"},
-    )
-    try:
-        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
-        username: str = payload.get("sub")
-        if username is None:
-            raise credentials_exception
-        token_data = TokenData(username=username)
-    except JWTError:
-        raise credentials_exception
-    user: Optional[User] = await user_crud.read(username=token_data.username)
-    if user is None:
-        raise credentials_exception
     return user
